requirements.txt
anthropic>=0.39.0
cryptography>=38.0.0
facebook-business>=20.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
sqlalchemy>=2.0.0
typer>=0.12.0
streamlit>=1.35.0
ruff>=0.5.0
pytest>=8.0.0
